Fixed bug where an aliased symbol was breaking DWARF filtering
On my machine (Debian amd64) I was seeing

 dima@shorty:$ ltrace -L -x '*ftell@libc.so*' ./dwarf
 ftell@libc.so.6(0x7f71126bc4e0)     = -1

which is corect; ftell() DWARF information was correctly interpreted, and we
know that ftell() takes one argument. However filtering more specifically, the
DWARF parsing appears to fail:

 dima@shorty:$ ltrace -L -x 'ftell@libc.so*' ./dwarf
 ftell@libc.so.6(0x7fb58872f4e0, 0x7fffb846b0c0, 7, 0x7fffb846b080)     = -1

Here we use the default (and wrong) prototype for ftell(). The issue is that
ftell() is an aliased symbol, and the DWARF parser only adds symbols that pass
the filter. This patch looks to see if the alias names pass the filter too, and
if they do, the DWARF prototypes are accepted. This makes the above work.