Detect return from rt_sigreturn syscalls
On MIPS, sigreturn/rt_sigreturn does not return control to the instruction
after the syscall instruction. Instead, it transfers control directly to
the point where the signal occurred. If this transfer is not correctly
marked as a SYSRET event, it gets mis-categorized as a breakpoint event
and generates "unexpected breakpoint" messages. The sigreturn
frame remains on the callstack until the next sigreturn, which is then
misinterpreted as as a return from the previous 'unfinished' syscall.

If the top of the callstack is a sigreturn or rt_sigreturn syscall, we
must force the current event to be treated as a SYSRET without checking
for a preceding SYSCALL instruction.

Verify by invoking:
$ ltrace [-S] openssl speed