Add a way to force a line to be interpreted as a function prototype
Since "import" is not a keyword in C, it might be used as a type name.
However, a function prototype with "import" as the return type would
be interpreted as an import directive. So provide a new keyword,
"function", that can be used to force a line to be interpreted as a
prototype.

Naturally, the new keyword will also work if the return type is
"function" or "typedef".